Style the background and add and style the "My To-Do List" heading of my-to-do-list.html
I added the royal blue background color and changed the font family and color of the text on the my-to-do-list.html page in order to ensure that the text on the page was visible and to alter the default color and font family of the webpage. I styled the heading by aligning the heading in the center of the webpage, increasing the text's font weight, and adding a text shadow.