Add and style form on my-to-do-list.html
I added the "to-do-list file" form with two inputs, a text field with the "to-do item" id and a button with an "add-button" id, so that users can input data into the form. I styled the text field by increasing its size and text size. I styled the button by changing the font family, increasing the size of the button, adding a white background color and black border, increasing the button's bottom margin, and ensuring that the text inside of the button appears bold. When the mouse is hovering over the button, the button's background color will appear light grey.