Add and style list of to-do items entered by users
I added JavaScript code that would enable users to type a to-do item into the text field and click the add button to add the to-do item to the newly-added unordered list on the webpage. I styled the unordered list by increasing the font size, aligning the text in the center of the webpage, adding a white background color, changing the text color to black, and adjusting the spacing of the list. I styled each individual list item by aligning the text to the left of the unordered list container, adding a border, placing each of the bullet points inside of the white background area, and adjusting the spacing of each list item.