Fail to allow users to add an empty string to the list
In the if-else statement that I added to the JavaScript code, I specified that nothing should occur if the input was trimmed of white space and an empty string remained. I also specified that a user should be able to add and remove a to-do list item if the string is not empty.