Add counter to the to-do-list
I added a count variable to my JavaScript code and added 1 to this variable when a new to-do item had been added to the list and subtracted 1 from this variable when a to-do item had been removed from the list. The newly-added heading with the "counter" id displayed the updated score to the users.